+++ CHANGELOG.md
@@ -2,6 +2,11 @@
 
 ### Libraries
 
+#### `dart:core`
+
+- `String.fromCharCodes` now allow `start` and `end` to be after the end of
+  the `Iterable` argument, just like `skip` and `take` does on an `Iterable`.
+
 #### `dart:nativewrappers`
 
 - **Breaking Change** [#51896][]: The NativeWrapperClasses are marked `base` so